# openapi.yaml
openapi: 3.0.3
...
info:
  title: Ryden REST API
  version: "1.0"
  description: >
    API REST (Jersey/JAX-RS) de Ryden — plataforma P2P de alquiler de autos.
    Reemplaza la capa Spring MVC + JSP. Consumida por una SPA servida como
    estáticos desde Tomcat (sin SSR). Esta especificación es el contrato
    público (DTOs), independiente del modelo de dominio.
servers:
  - url: /webapp
    description: >
      Context path en Tomcat local (WAR desplegado como webapp.war).
      En Jetty dev (`mvn jetty:run`) el context path es `/` — ajustar el prefijo al probar.
  - url: /
    description: Raíz de la API en Jetty dev o despliegue sin context path
...
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [users]
      summary: Obtener un usuario (perfil público o privado)
      description: >
        URN canónica del usuario. La representación depende del header Accept
        (dos schemas fijos, no recorte por permisos):
        - `application/vnd.paw.user.v1+json` — perfil público (sin email, CBU, teléfono, role, blocked).
        - `application/vnd.paw.user.private.v1+json` — vista self/admin (403 si no autorizado).
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
            example: application/vnd.paw.user.v1+json
      responses:
        "200":
          description: Usuario.
          content:
            application/vnd.paw.user.v1+json:
              schema: { $ref: "#/components/schemas/UserDto" }
            application/vnd.paw.user.private.v1+json:
              schema: { $ref: "#/components/schemas/UserPrivateDto" }
            application/vnd.paw.user.v1+xml:
              schema: { $ref: "#/components/schemas/UserDto" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [users]
      summary: Reemplazar el perfil completo del usuario (self/admin)
      description: >
        PUT reemplaza la entidad completa (atributos null se interpretan como
        borrado del dato). Para cambios parciales/transiciones usar PATCH.
        Los `links` NO se persisten, así que enviarlos o no es indistinto.
      requestBody:
        required: true
        content:
          application/vnd.paw.user.v1+json:
            schema: { $ref: "#/components/schemas/UserDto" }
      responses:
        "200":
          description: Usuario reemplazado.
          content:
            application/vnd.paw.user.v1+json:
              schema: { $ref: "#/components/schemas/UserDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [users]
      summary: Actualización parcial / transiciones de estado del usuario
      description: >
        Un único endpoint PATCH absorbe TODAS las "acciones" del MVC viejo
        que eran cambios de estado sobre el usuario. El cuerpo solo lleva los
        atributos a cambiar; el resource autoriza según quién sea el llamante:

        - Actualizar perfil (nombre, teléfono, fecha nac., about, cbu, locale): self (Bearer).
        - Cambiar password (logueado) → { "password": "...", "currentPassword": "..." }.
        - Resetear password olvidado → { "password": "...", "passwordConfirm": "..." }
          autenticado con Basic email:resetOtp (authority ROLE_PASSWORD_RESET_OTP).
          El OTP se consume al completar el reset (decisión D4).
        - Promover/degradar admin → { "role": "admin" | "user" } (solo admin).
        - Bloquear/desbloquear → { "blocked": true | false } (solo admin).
        - Validar documentos/identidad → { "identityValidated": true } (solo admin).

        La verificación de email NO usa PATCH: el OTP se presenta como
        Basic email:otp en cualquier request autenticado (p. ej. login);
        el backend consume el código y marca emailVerified (decisión D3).

        Reglas de negocio (no auto-bloqueo, no bloquear al grantor, etc.) se
        mapean a 403 vía ExceptionMapper.
      requestBody:
        required: true
        content:
          application/vnd.paw.user.v1+json:
            schema: { $ref: "#/components/schemas/UserPatchDto" }
      responses:
        "200":
          description: Usuario actualizado.
          content:
            application/vnd.paw.user.v1+json:
              schema: { $ref: "#/components/schemas/UserDto" }
            application/vnd.paw.user.private.v1+json:
              schema: { $ref: "#/components/schemas/UserPrivateDto" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Eliminar usuario (self/admin)
      responses:
        "204": { description: Eliminado. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # --- Imagen de perfil (sub-recurso binario débil) ---
  /users/{id}/profile-picture:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [users, files]
      summary: Foto de perfil (binario público)
      responses:
        "200": { $ref: "#/components/responses/BinaryImage" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [users, files]
      summary: Subir/reemplazar foto de perfil (self)
      requestBody:
        required: true
        content:
          image/*:
            schema: { type: string, format: binary }
      responses:
        "204": { description: Subida. }
        "400": { $ref: "#/components/responses/ValidationError" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [users, files]
      summary: Borrar foto de perfil (self)
      responses:
        "204": { description: Borrada. }
        "403": { $ref: "#/components/responses/Forbidden" }

  # --- Documentos de perfil (licencia / identidad) ---
...
  /users/{id}/favorites:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [users]
      summary: Autos favoritos del usuario (self) — colección paginada
      description: Los miembros son autos; se hipervinculan, no se embeben.
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Lista de autos favoritos.
          headers:
            Link: { $ref: "#/components/headers/Link-Pagination" }
            X-Total-Count: { $ref: "#/components/headers/X-Total-Count" }
          content:
            application/vnd.paw.car.v1+json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CarDto" }
        "204": { description: Sin favoritos. }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users/{id}/favorites/{carId}:
    parameters:
      - $ref: "#/components/parameters/idPath"
      - name: carId
        in: path
        required: true
        schema: { type: integer, format: int64 }
    put:
      tags: [users]
      summary: Marcar un auto como favorito (idempotente)
      description: >
        Reemplaza el "toggle" del MVC por operaciones idempotentes:
        PUT agrega (favoritear), DELETE quita. El cliente ya sabe el estado.
      responses:
        "204": { description: Favorito agregado (o ya existía). }
        "403":
          description: No se puede favoritear el auto propio.
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [users]
      summary: Quitar de favoritos (idempotente)
      responses:
        "204": { description: Quitado (o no estaba). }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users/{id}/reviews:
    parameters:
      - $ref: "#/components/parameters/idPath"
    get:
      tags: [users, reviews]
      summary: Reseñas recibidas por el usuario (perfil público)
      description: >
        Reseñas que el usuario RECIBIÓ, combinando las recibidas como dueño
        (de conductores) y como conductor (de dueños) en una sola lista, más
        recientes primero. Proyección de solo lectura: trae el nombre del autor,
        rating, fecha y comentario (sin id ni datos sensibles). Pública.
      security:
        - {}
      responses:
        "200":
          description: Lista de reseñas recibidas.
          content:
            application/vnd.paw.review.v1+json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserReviewDto" }
        "204": { description: Sin reseñas. }
        "404": { $ref: "#/components/responses/NotFound" }

  #############################################################################
  # CREDENCIALES (OTP temporales) — sub-recursos sin verbos en la URL
  #############################################################################
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >
        Authorization: Basic base64(email:secret) a CUALQUIER endpoint.
        El secret depende del contexto:
        - Cuenta verificada: password (BCrypt).
        - Verificación de email (cuenta sin verificar): OTP de 6 dígitos enviado
          por mail tras POST /users. Se consume al autenticar; devuelve JWT.
        - Reset de password olvidado: OTP enviado tras POST /credentials. Otorga
          authority ROLE_PASSWORD_RESET_OTP; se consume al completar PATCH
          /users/{id} con password + passwordConfirm.
        Tras auth válida el server responde con X-Access-Token, X-Refresh-Token
        y Link rel="authenticated-user". No hay /login ni /verify.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Requests normales: Authorization: Bearer <access_token>. Cuando vence
        (401), se manda el Refresh Token (también como Bearer) a cualquier
        endpoint y el server devuelve un par nuevo en headers. No hay /refresh:
        el JWT es autocontenido y el server distingue access de refresh.
  parameters:
    idPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (la metadata de paginación va en el header Link, no en el body).
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 12 }
  headers:
...
    Link-Pagination:
      description: >
        Links de paginación con rel="first" | "prev" | "next" | "last",
        honrando los query params de filtro activos. El cliente los navega
        genéricamente; NO construye URLs a mano.
      schema: { type: string }
      example: '</cars?page=2&category=suv>; rel="next", </cars?page=5&category=suv>; rel="last"'
    X-Total-Count:
      description: Cantidad total de elementos (opcional, complementa Link).
      schema: { type: integer }
...
  responses:
    Unauthorized:
      description: 401 — Falta autenticación o token inválido/vencido.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
    Forbidden:
      description: 403 — Autenticado pero sin permiso.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
    NotFound:
      description: 404 — El recurso no existe.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ErrorDto" }
...
    ValidationError:
      description: 400 — Request mal formado o validación fallida.
      content:
        application/vnd.paw.error.v1+json:
          schema: { $ref: "#/components/schemas/ValidationErrorDto" }
    BinaryImage:
      description: Bytes de una imagen (content-type detectado).
      content:
        image/*:
          schema: { type: string, format: binary }
...
  schemas:
    Links:
      type: object
      description: Bloque de hipervínculos generado on-the-fly. Siempre incluye self. NUNCA se persiste.
      additionalProperties: { type: string, format: uri }
      properties:
        self: { type: string, format: uri }
      required: [self]
...
    UserDto:
      type: object
      description: >
        Representación pública fija (`Accept: application/vnd.paw.user.v1+json`).
        NO incluye email, CBU, teléfono, role ni blocked — usar UserPrivateDto para eso.
      properties:
        forename: { type: string }
        surname: { type: string }
        memberSince: { type: string, format: date }
        emailVerified: { type: boolean }
        licenseValidated: { type: boolean }
        identityValidated: { type: boolean }
        licenseUploaded: { type: boolean }
        identityUploaded: { type: boolean }
        ratingAsRider: { type: number, nullable: true }
        ratingAsOwner: { type: number, nullable: true }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self, profilePicture, cars, reservations, favorites, documents, reviews.
    UserPrivateDto:
      type: object
      description: >
        Representación privada fija (`Accept: application/vnd.paw.user.private.v1+json`).
        Solo self/admin; 403 si se pide sin permiso.
      properties:
        forename: { type: string }
        surname: { type: string }
        email: { type: string, format: email }
        phoneNumber: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        about: { type: string, nullable: true }
        cbu: { type: string, nullable: true }
        memberSince: { type: string, format: date }
        latestLocale: { type: string, nullable: true }
        emailVerified: { type: boolean }
        licenseValidated: { type: boolean }
        identityValidated: { type: boolean }
        licenseUploaded: { type: boolean }
        identityUploaded: { type: boolean }
        blocked: { type: boolean }
        role: { type: string, enum: [user, admin] }
        ratingAsRider: { type: number, nullable: true }
        ratingAsOwner: { type: number, nullable: true }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self, profilePicture, cars, reservations, favorites, documents, reviews.
...
    UserPatchDto:
      type: object
      description: >
        Todos opcionales; solo se envían los atributos a cambiar. OTPs de
        verificación/reset NO van en el body — viajan en Authorization Basic
        email:otp. La autorización depende de cada campo (self vs admin) y de
        la credencial presentada.
      properties:
        forename: { type: string }
        surname: { type: string }
        phoneNumber: { type: string }
        birthDate: { type: string, format: date }
        about: { type: string }
        cbu: { type: string }
        latestLocale: { type: string }
        password: { type: string, format: password, writeOnly: true }
        passwordConfirm: { type: string, format: password, writeOnly: true, description: Requerido junto con password en reset olvidado. }
        currentPassword: { type: string, format: password, writeOnly: true, description: Requerido al cambiar password estando logueado. }
        role: { type: string, enum: [user, admin], description: Solo admin. }
        blocked: { type: boolean, description: Solo admin. }
        identityValidated: { type: boolean, description: Solo admin. }
        licenseValidated: { type: boolean, description: Solo admin. }

    # ----- CAR -----
    CarType:
      type: string
      enum: [sedan, hatchback, suv, coupe, convertible, wagon, van, pickup]
    Powertrain:
      type: string
      enum: [gasoline, diesel, electric, hybrid, cng]
    Transmission:
      type: string
      enum: [manual, automatic, semi_automatic]
    CarStatus:
      type: string
      enum: [active, paused, admin_paused, lack_doc, unavailable, deactivated]
    PriceMarketPosition:
      type: string
      enum: [below_market, at_market, above_market]
      description: Consumer browse card badge relative to same brand/model market average.
    CarDto:
      type: object
      properties:
        plate: { type: string }
        year: { type: integer, nullable: true }
        powertrain: { $ref: "#/components/schemas/Powertrain" }
        transmission: { $ref: "#/components/schemas/Transmission" }
        type: { $ref: "#/components/schemas/CarType" }
        status: { $ref: "#/components/schemas/CarStatus" }
        description: { type: string, nullable: true }
        minimumRentalDays: { type: integer }
        ratingAvg: { type: number, nullable: true }
        dayPrice:
          type: number
          nullable: true
          description: Lowest active day price; populated on browse card projections only.
        brandName: { type: string }
        modelName: { type: string }
        modelValidated: { type: boolean }
        hasInsurance: { type: boolean }
        priceMarketPositionModifier:
          allOf:
            - $ref: "#/components/schemas/PriceMarketPosition"
          nullable: true
          description: Consumer browse/favorites cards when comparable market data exists.
        marketAveragePrice:
          type: number
          nullable: true
          description: Same brand/model market average day price backing the badge.
        marketSampleCount:
          type: integer
          format: int64
          nullable: true
          description: Number of comparable active listings used for the badge.
        createdAt: { type: string, format: date-time }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self, owner, model, brand, pictures, availabilities, insurance, reviews.
...
    UserReviewDto:
      type: object
      description: >
        Reseña RECIBIDA por un usuario (perfil público). Proyección de solo
        lectura desde ReviewItemDto; no expone el id en el cuerpo, pero
        links.self ya es la URN única de la reseña (mismo recurso que ReviewDto).
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        authorName: { type: string, description: Nombre del que dejó la reseña. }
        createdAt: { type: string, format: date, description: Fecha local de la reseña. }
        links:
          allOf:
            - $ref: "#/components/schemas/Links"
          description: self (único, /reservations/{id}/reviews/{reviewId}).

    # ----- CATALOG -----
...
    ErrorDto:
      type: object
      properties:
        status: { type: integer }
        code: { type: string }
        message: { type: string }
    ValidationErrorDto:
      type: object
      properties:
        status: { type: integer, example: 400 }
        message: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }